pyproject.toml
[tool.poetry]
name = "google-scholar"
version = "0.1.0"
description = ""
authors = ["ritikrajdev <ritikrajdev761@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.6"
requests = "^2.26.0"
beautifulsoup4 = "^4.9.3"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
ipython = "^7.26.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
